move react hooks from css engine to shared styles, it's too tight with app
