+++ packages/css-engine/src/index.ts
@@ -1,2 +1,1 @@
 export * from "./core";
-export * from "./react";
